fixup freetype issue, so that it links to our libzlib.dll
the freetype configure script seems buggy somehow... it seems to
produce a libfreetype that always requires zlib1.dll whenever we
install it, even if our zlib instance is not called that. it may
be that `make install` doesn't really work when you are
cross-compiling... or libtool may be to blame somehow.

what we do now is, make install in order to get the headers, then
remove all the binaries and copy the seemingly-correct dll from
objs/.libs in the freetype folder...